done all

+++ lib/src/model/all_data.dart
@@ -9,7 +9,7 @@ class AllData {
   int? active;
   int? critical;
   int? casesPerOneMillion;
-  double? deathsPerOneMillion;
+  int? deathsPerOneMillion;
   int? tests;
   double? testsPerOneMillion;
   int? population;
